Add test for lookup function
Changed the file naming scheme to be simpler in anticipation for using
file names to open files instead of a slice of pointers to files.
Changed the hash index to store string-string pairs instead of string-
int pairs. Changed loops to be range loops. Tests not passing. EOF error
due to trying to read elemSize bytes from the data file. Going to change
scheme to be all one line with each entry delineated by number of bytes
as a entry header.